Create a common base class for topic-based sinks.
